fix: payment api call

+++ app/src/main/java/com/trekkstay/hotel/feature/reservation/data/models/payment_model.kt
@@ -64,10 +64,9 @@ data class PaymentModel(
         }
 
         fun fromMap(map: DataMap): PaymentModel {
-
             return PaymentModel(
                 reservationId = map["reservation_id"] as String,
-                amount = map["amount"] as Int,
+                amount = (map["amount"] as Double).toInt(),
                 userId = map["user_id"] as String,
                 method = map["method"] as String,
                 status = map["status"] as String
